slice and move AutomatonGenerator
